Add whitespace examples.
