+++ analyzers/tests/SonarAnalyzer.Test/TestCases/RegularExpressions/RegexShouldNotContainMultipleSpaces.vb
